% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psuedo_MLE_clayton.R
\name{est_cop_par}
\alias{est_cop_par}
\title{TODO: short description}
\usage{
est_cop_par(copula = "clayton", data, niter = 50)
}
\arguments{
\item{copula}{string indicating the copula kind.}

\item{data}{input data.}

\item{niter}{number of iterations to run.}
}
\value{

}
\description{
Longer description
}
