% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_clayton.R
\name{npest.star_12}
\alias{npest.star_12}
\title{Function to return estimated SD, S1star, S2star}
\usage{
npest.star_12(X)
}
\description{
Function to return estimated SD, S1star, S2star
}
\keyword{internal}
