% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{TODO: Test data generated from XXXX..}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
\item{price}{price, in US dollars}
\item{carat}{weight of the diamond, in carats}
...
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
test_data
}
\description{
Longer description of the data.
}
\keyword{datasets}
